+++ packages/backend/package.json
@@ -14,7 +14,7 @@
     "test:integration": "bun test src/tests/integration/",
     "test:e2e": "bun test src/tests/e2e/",
     "test:all": "bun test",
-    "test:sqlite": "npx vitest run src/tests/node/sqlite-repositories.vitest.ts",
+    "test:sqlite": "bun test src/tests/integration/sqlite-repositories.test.ts",
     "db:generate": "drizzle-kit generate",
     "db:migrate": "bun run src/db/migrate.ts",
     "db:studio": "drizzle-kit studio"
@@ -24,7 +24,6 @@
     "@hono/zod-validator": "^0.7.6",
     "@simplewebauthn/server": "^13.3.0",
     "@simplewebauthn/types": "^12.0.0",
-    "better-sqlite3": "^12.10.0",
     "blurhash": "^2.0.5",
     "bullmq": "^5.76.8",
     "drizzle-orm": "^0.45.2",
@@ -43,7 +42,6 @@
     "zod": "^4.4.3"
   },
   "devDependencies": {
-    "@types/better-sqlite3": "^7.6.13",
     "@types/eventsource": "^3.0.0",
     "@types/pg": "^8.20.0",
     "@types/pino": "^7.0.5",
@@ -51,7 +49,6 @@
     "@types/web-push": "^3.6.4",
     "bun-types": "^1.3.14",
     "drizzle-kit": "^0.31.10",
-    "typescript": "^5.9.3",
-    "vitest": "^4.1.6"
+    "typescript": "^5.9.3"
   }
 }
