+++ packages/backend/src/repositories/sqlite/SqliteCustomEmojiRepository.ts
@@ -7,7 +7,7 @@
  */
 
 import { eq, and, isNull, like, inArray, sql } from "drizzle-orm";
-import type { BetterSQLite3Database } from "drizzle-orm/better-sqlite3";
+import type { BunSQLiteDatabase } from "drizzle-orm/bun-sqlite";
 import { customEmojis, type CustomEmoji, type NewCustomEmoji } from "../../db/schema/sqlite.js";
 import type * as sqliteSchema from "../../db/schema/sqlite.js";
 import type {
@@ -15,7 +15,7 @@ import type {
   ListCustomEmojisOptions,
 } from "../../interfaces/repositories/ICustomEmojiRepository.js";
 
-type SqliteDatabase = BetterSQLite3Database<typeof sqliteSchema>;
+type SqliteDatabase = BunSQLiteDatabase<typeof sqliteSchema>;
 
 /**
  * SQLite implementation of Custom Emoji Repository
