+++ packages/backend/src/repositories/sqlite/SqliteDriveFolderRepository.ts
@@ -7,13 +7,13 @@
  */
 
 import { eq, and, sql, isNull, desc } from "drizzle-orm";
-import type { BetterSQLite3Database } from "drizzle-orm/better-sqlite3";
+import type { BunSQLiteDatabase } from "drizzle-orm/bun-sqlite";
 import { driveFolders } from "../../db/schema/sqlite.js";
 import type * as sqliteSchema from "../../db/schema/sqlite.js";
 import type { IDriveFolderRepository } from "../../interfaces/repositories/IDriveFolderRepository.js";
 import type { DriveFolder } from "shared";
 
-type SqliteDatabase = BetterSQLite3Database<typeof sqliteSchema>;
+type SqliteDatabase = BunSQLiteDatabase<typeof sqliteSchema>;
 
 /**
  * SQLite implementation of drive folder repository
