+++ packages/backend/src/repositories/sqlite/SqliteFollowRepository.ts
@@ -7,13 +7,13 @@
  */
 
 import { eq, and, or, sql } from "drizzle-orm";
-import type { BetterSQLite3Database } from "drizzle-orm/better-sqlite3";
+import type { BunSQLiteDatabase } from "drizzle-orm/bun-sqlite";
 import { follows, users } from "../../db/schema/sqlite.js";
 import type * as sqliteSchema from "../../db/schema/sqlite.js";
 import type { IFollowRepository } from "../../interfaces/repositories/IFollowRepository.js";
 import type { Follow } from "shared";
 
-type SqliteDatabase = BetterSQLite3Database<typeof sqliteSchema>;
+type SqliteDatabase = BunSQLiteDatabase<typeof sqliteSchema>;
 
 /**
  * SQLite implementation of follow repository
