+++ packages/backend/src/repositories/sqlite/SqliteModerationAuditLogRepository.ts
@@ -7,7 +7,7 @@
  */
 
 import { eq, sql, desc, and } from "drizzle-orm";
-import type { BetterSQLite3Database } from "drizzle-orm/better-sqlite3";
+import type { BunSQLiteDatabase } from "drizzle-orm/bun-sqlite";
 import { moderationAuditLogs, type ModerationAuditLog } from "../../db/schema/sqlite.js";
 import type * as sqliteSchema from "../../db/schema/sqlite.js";
 import type {
@@ -17,7 +17,7 @@ import type {
 } from "../../interfaces/repositories/IModerationAuditLogRepository.js";
 import { generateId } from "shared";
 
-type SqliteDatabase = BetterSQLite3Database<typeof sqliteSchema>;
+type SqliteDatabase = BunSQLiteDatabase<typeof sqliteSchema>;
 
 /**
  * SQLite implementation of Moderation Audit Log Repository
