+++ packages/backend/src/repositories/sqlite/SqlitePasskeyChallengeRepository.ts
@@ -7,7 +7,7 @@
  */
 
 import { eq, lt, sql } from "drizzle-orm";
-import type { BetterSQLite3Database } from "drizzle-orm/better-sqlite3";
+import type { BunSQLiteDatabase } from "drizzle-orm/bun-sqlite";
 import {
   passkeyChallenges,
   type PasskeyChallenge,
@@ -16,7 +16,7 @@ import {
 import type * as sqliteSchema from "../../db/schema/sqlite.js";
 import type { IPasskeyChallengeRepository } from "../../interfaces/repositories/IPasskeyChallengeRepository.js";
 
-type SqliteDatabase = BetterSQLite3Database<typeof sqliteSchema>;
+type SqliteDatabase = BunSQLiteDatabase<typeof sqliteSchema>;
 
 /**
  * SQLite implementation of Passkey Challenge Repository
