+++ packages/backend/src/repositories/sqlite/SqliteReactionRepository.ts
@@ -7,13 +7,13 @@
  */
 
 import { eq, and, sql, desc, inArray } from "drizzle-orm";
-import type { BetterSQLite3Database } from "drizzle-orm/better-sqlite3";
+import type { BunSQLiteDatabase } from "drizzle-orm/bun-sqlite";
 import { reactions } from "../../db/schema/sqlite.js";
 import type * as sqliteSchema from "../../db/schema/sqlite.js";
 import type { IReactionRepository } from "../../interfaces/repositories/IReactionRepository.js";
 import type { Reaction } from "shared";
 
-type SqliteDatabase = BetterSQLite3Database<typeof sqliteSchema>;
+type SqliteDatabase = BunSQLiteDatabase<typeof sqliteSchema>;
 
 /**
  * SQLite implementation of reaction repository
