+++ packages/backend/src/repositories/sqlite/SqliteRoleAssignmentRepository.ts
@@ -7,7 +7,7 @@
  */
 
 import { eq, and, sql, lt } from "drizzle-orm";
-import type { BetterSQLite3Database } from "drizzle-orm/better-sqlite3";
+import type { BunSQLiteDatabase } from "drizzle-orm/bun-sqlite";
 import { roleAssignments, roles, type Role, type RoleAssignment } from "../../db/schema/sqlite.js";
 import type * as sqliteSchema from "../../db/schema/sqlite.js";
 import type {
@@ -16,7 +16,7 @@ import type {
 } from "../../interfaces/repositories/IRoleAssignmentRepository.js";
 import { generateId } from "shared";
 
-type SqliteDatabase = BetterSQLite3Database<typeof sqliteSchema>;
+type SqliteDatabase = BunSQLiteDatabase<typeof sqliteSchema>;
 
 /**
  * SQLite implementation of Role Assignment Repository
