+++ packages/backend/src/repositories/sqlite/SqliteRoleRepository.ts
@@ -7,7 +7,7 @@
  */
 
 import { eq, asc, sql, desc } from "drizzle-orm";
-import type { BetterSQLite3Database } from "drizzle-orm/better-sqlite3";
+import type { BunSQLiteDatabase } from "drizzle-orm/bun-sqlite";
 import { roles, type Role } from "../../db/schema/sqlite.js";
 import type * as sqliteSchema from "../../db/schema/sqlite.js";
 import type {
@@ -17,7 +17,7 @@ import type {
 } from "../../interfaces/repositories/IRoleRepository.js";
 import { generateId } from "shared";
 
-type SqliteDatabase = BetterSQLite3Database<typeof sqliteSchema>;
+type SqliteDatabase = BunSQLiteDatabase<typeof sqliteSchema>;
 
 /**
  * SQLite implementation of Role Repository
