+++ packages/backend/src/repositories/sqlite/SqliteUserWarningRepository.ts
@@ -7,13 +7,13 @@
  */
 
 import { eq, and, desc, isNull, sql, or } from "drizzle-orm";
-import type { BetterSQLite3Database } from "drizzle-orm/better-sqlite3";
+import type { BunSQLiteDatabase } from "drizzle-orm/bun-sqlite";
 import { userWarnings, type UserWarning } from "../../db/schema/sqlite.js";
 import type * as sqliteSchema from "../../db/schema/sqlite.js";
 import type { IUserWarningRepository } from "../../interfaces/repositories/IUserWarningRepository.js";
 import { generateId } from "shared";
 
-type SqliteDatabase = BetterSQLite3Database<typeof sqliteSchema>;
+type SqliteDatabase = BunSQLiteDatabase<typeof sqliteSchema>;
 
 /**
  * SQLite implementation of user warning repository
